requirements.txt
streamlit
numpy
pandas
pywavelets
plotly
scipy
websocket-client
threading
json